10x speedup of proportion_of_energy_submetered ;)
Prior to this change, the function took 154 seconds for REDD house 1
on my 2.4GHz i5 M450 HP laptop.

It now takes 16 seconds.

The change was simple: profiling revealed that
`pytables.select_coords` was the bottleneck.  We only use this
function to get the `look_ahead`.  So this patch changes the behaviour
of `DataStore.load` so it defaults to *not* load `look_ahead`.

(This change breaks a test.  Will fix that...)