I think this solves issue #133.  Appears to make it slow though.
The problem was caused by the fact that
we cannot simply test if len(data) == chunksize
because store.select(chunksize=chunksize) doesn't
appear to respect the chunksize argument!

TODO: report bug to Pandas / PyTables

Alternative approach: before this loop,
make a copy of the generator using tee()
and loop through it to find out how many
items there are in it, then count through those
in this loop.