Fix issue with ElecMeter room.  Fixes #405.
(in an early draft of NILM Metadata, `ElecMeter:room` was defined as an object.  e.g. in the metadata you'd have something like `room: {'name': 'bedroom', 'instance': 2}`.  So NILMTK tried to do `room['name']` to get 'bedroom'.  But then I simplified the metadata so rooms were specified as '<name>[,<instance>]` e.g. 'bedroom'.  So now `room` is a string and so Python was complaining that we were trying to index into a string with another string.  Doh.  Sorry!