Update of refit converter for clean data
Clean data has been published by the University of Strathclyde.

This does not have duplicate indices, but includes additional columns and a header row.

The file prefix has changed to Clean_House

- A datetime and unix datatime is included.
- An issues column to identify rows where the power of meters exceeds the limit for the meter =1, 0 otherwise

Issues are now dropped by the converter and the proportion dropped is reported as a proportion of the overall file.

Nan values are also checked for and reported, but not removed.

A couple of helpers to identify the prefix as well as the localisation datetime format.